chore: refactor auth guard
