+++ apps/drec-api/src/guards/index.ts
@@ -2,4 +2,4 @@ export * from './ActiveUserGuard';
 export * from './RolesGuard';
 export * from './PermissionGuard';
 export * from './WithoutAuthGuard';
-export * from './VerifiedUserGuard';
+export * from './AuthVerifiedGuard';
